Update Docker configuration and documentation
The Docker image used in the `srf-api-prod-docker-compose.yml` file has been updated to use a new image. In the documentation (`docker-prod.md`), the bash command for building and pushing the Docker image has been updated to reflect the new version tag. This ensures consistency and correct instructions for future Docker builds.